feat: enhance search functionality and UI feedback
Implemented a dropdown in App.vue for selecting which database table to search in. Added search term input field and magnify icon for initiating the search. Refined the Details dialog for a more user-friendly display of search results. Introduced a snackbar component for user notifications and feedback. Enhanced the preload.js to include IPC communication for table and column listings. Updated background.js to handle dynamic queries based on selected table and to only list FTS5 tables.

Closes #2